requirements.txt
alembic==1.0.8
Click==7.0
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.8
MarkupSafe==1.1.1
pkg-resources==0.0.0
psycopg2-binary==2.8.1
python-dateutil==2.8.0
python-dotenv==0.10.1
python-editor==1.0.4
six==1.12.0
SQLAlchemy==1.3.2
Werkzeug==0.15.2
WTForms==2.2.1
